feat(editor): Show actions in focus panel when multiple nodes are selected (no-changelog) (#18984)
