+++ packages/frontend/@n8n/design-system/src/components/N8nActionDropdown/ActionDropdown.stories.ts
@@ -30,7 +30,9 @@ const template: StoryFn = (args, { argTypes }) => ({
 	setup: () => ({ args }),
 	props: Object.keys(argTypes),
 	components: {
-		N8nActionDropdown,
+		// "as unknown ..." is a workaround for generic components.
+		// See https://github.com/storybookjs/storybook/issues/24238
+		N8nActionDropdown: N8nActionDropdown as unknown as Record<string, unknown>,
 	},
 	template: '<n8n-action-dropdown v-bind="args" />',
 });
