+++ packages/frontend/@n8n/design-system/src/types/action-dropdown.ts
@@ -2,8 +2,8 @@ import type { KeyboardShortcut } from '@n8n/design-system/types/keyboardshortcut
 
 import type { IconName } from '../components/N8nIcon/icons';
 
-export interface ActionDropdownItem {
-	id: string;
+export interface ActionDropdownItem<T extends string> {
+	id: T;
 	label: string;
 	badge?: string;
 	badgeProps?: Record<string, unknown>;
