+++ packages/frontend/editor-ui/src/components/canvas/experimental/components/ExperimentalNodeDetailsDrawer.test.ts
@@ -8,6 +8,7 @@ import { useWorkflowsStore } from '@/stores/workflows.store';
 import { createTestingPinia } from '@pinia/testing';
 import ExperimentalNodeDetailsDrawer from './ExperimentalNodeDetailsDrawer.vue';
 import { nextTick } from 'vue';
+import { fireEvent } from '@testing-library/vue';
 
 const renderComponent = createComponentRenderer(ExperimentalNodeDetailsDrawer);
 
@@ -62,7 +63,7 @@ describe('ExperimentalNodeDetailsDrawer', () => {
 			pinia,
 			props: {
 				node: mockNodes[0],
-				nodes: [mockNodes[0]],
+				nodeIds: ['node1'],
 			},
 		});
 
@@ -76,5 +77,43 @@ describe('ExperimentalNodeDetailsDrawer', () => {
 		await nextTick();
 
 		await rendered.findByDisplayValue('after update');
+	});
+
+	describe('when multiple nodes are selected', () => {
+		it('should show the number of selected nodes and available actions', () => {
+			const rendered = renderComponent({
+				pinia,
+				props: {
+					node: mockNodes[0],
+					nodeIds: ['node1', 'node2'],
+				},
+			});
+
+			expect(rendered.getByText('2 nodes selected')).toBeInTheDocument();
+
+			const buttons = rendered.getAllByRole('button');
+			expect(buttons.length).toBeGreaterThan(0);
+
+			expect(rendered.getByText('Copy 2 nodes')).toBeInTheDocument();
+			expect(rendered.getByText('Duplicate 2 nodes')).toBeInTheDocument();
+			expect(rendered.getByText('Delete 2 nodes')).toBeInTheDocument();
+		});
+
+		it('should emit contextMenuAction event when a button is pressed', async () => {
+			const rendered = renderComponent({
+				pinia,
+				props: {
+					node: mockNodes[0],
+					nodeIds: ['node1', 'node2'],
+				},
+			});
+
+			const copyButton = rendered.getByText('Copy 2 nodes').closest('button')!;
+
+			await fireEvent.click(copyButton);
+
+			expect(rendered.emitted('contextMenuAction')).toBeTruthy();
+			expect(rendered.emitted('contextMenuAction')?.[0]).toEqual(['copy', ['node1', 'node2']]);
+		});
 	});
 });
