+++ packages/frontend/editor-ui/src/features/dataStore/components/dataGrid/ColumnHeader.vue
@@ -20,14 +20,14 @@ const i18n = useI18n();
 
 const isHovered = ref(false);
 const isDropdownOpen = ref(false);
-const dropdownRef = ref<InstanceType<typeof N8nActionDropdown>>();
+const dropdownRef = ref<{ $el?: Node }>();
 
 const enum ItemAction {
 	Delete = 'delete',
 }
 
-const onItemClick = (action: string) => {
-	if (action === (ItemAction.Delete as string)) {
+const onItemClick = (action: ItemAction) => {
+	if (action === ItemAction.Delete) {
 		props.params.onDelete(props.params.column.getColId());
 	}
 };
